Add ci job to check if nightly succeeded (#45509)
Summary:
This change adds a job that runs 2 hours after the nightlies. This jobs returns successfully if the nightly has been published or it report an error in case it has failed.

We will hook this signal to the internal system to be notified about Nightlies failures

## Changelog:

[Internal] - Add jobs to check for nightlies

Pull Request resolved: https://github.com/facebook/react-native/pull/45509

Test Plan: Test the Action on GHA

Reviewed By: blakef

Differential Revision: D59907442

Pulled By: cipolleschi

fbshipit-source-id: 3b35aa2ad69b376c65a765f740a1d6e6ed8ad99f