fix Android's Suspense issue when used with useTransition, startTransition and useDeferredValue (#45576)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/45576

changelog: [internal]

The problem:

Android's mounting layer is architected in a way where React can only go forward in revisions. For example, it can't go to an older revision of a shadow node. The updates applied on Android won't be correct.

So going from revision 1 to revision 2 on Android is ok. Going from revision 2, back to revision 1 won't work. This behaviour is exactly what React will do with Suspense and when state update is wrapped in a transition.

The ideal solution would be to use entire Prop struct to do props diffing, like on [iOS](https://fburl.com/code/6v6tzyt5). Before we do that and re-write a large part of Android, we can add a special case to handle `display: 'none'`, which is the style that Suspense uses to [hide and unhide](https://github.com/facebook/react/blob/f603426f917314561c4289734f39b972be3814af/packages/react-native-renderer/src/ReactFiberConfigFabric.js#L433) the boundary. In this diff, before passing down `rawProps`, which is the change set coming React, we can inspect to detect the faulty case and correct it.

Reviewed By: javache

Differential Revision: D60036222