Add "enableBackgroundStyleApplicator" feature flag
Summary:
Going to use this to gate usage of the applicator, for each of the components, including `<View>`, which

This also conveniently sidesteps some unsavory reflection on View member happening that we can't clean up yet.

I defaulted this one to on by default, since it shouldn't be major behavior change compared to flows already using CSSBackgroundDrawable for rendering. Images must still pass flag to not use Fresco based rendering.

Changelog: [Internal]

Differential Revision: D60265327