Add more style value types (#45687)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/45687

This adds some more enums and data classes to encapsulate style values we are working with for border/background rendering. Right now, a lot of these are passed around as strings, or raw ints (of differing ordinals). These will be used as the public API of `BackgroundStyleApplicator` up the stack.

Changelog: [Internal]

Reviewed By: rshest

Differential Revision: D60252277

fbshipit-source-id: 0f8001869421ffffae9727c7904bf5e395505c08