Fix crash from boxShadow conditional canvas.save()
Summary:
D59300215 noticed that the drawable was leaking a clipping rect for the rest of the operations, and added a `save/restore` pair, but the save happens conditionally, so we can restore more often than we save.

Scope the save/restore, to just the area setting context then drawing onto the canvas.

```
java.lang.IllegalStateException: Underflow in restore - more restores than saves
                                                                                                    	at android.graphics.Canvas.restore(Canvas.java:647)
                                                                                                    	at com.facebook.react.uimanager.drawable.OutsetBoxShadowDrawable.draw(OutsetBoxShadowDrawable.kt:110)
                                                                                                    	at android.graphics.drawable.LayerDrawable.draw(LayerDrawable.java:1019)
```

Differential Revision: D60375357